prepare to handle the type argument in `array()`
