+++ r/R/RcppExports.R
@@ -1,10 +1,6 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
-Array__from_vector <- function(x) {
-    .Call(`_arrow_Array__from_vector`, x)
-}
-
 Array__Slice1 <- function(array, offset) {
     .Call(`_arrow_Array__Slice1`, array, offset)
 }
@@ -87,6 +83,10 @@ RecordBatch__to_dataframe <- function(batch, use_threads) {
 
 Table__to_dataframe <- function(table, use_threads) {
     .Call(`_arrow_Table__to_dataframe`, table, use_threads)
+}
+
+Array__from_vector <- function(x, type) {
+    .Call(`_arrow_Array__from_vector`, x, type)
 }
 
 ArrayData__get_type <- function(x) {
