+++ r/man/array.Rd
@@ -4,12 +4,12 @@
 \alias{array}
 \title{create an \link[=arrow__Array]{arrow::Array} from an R vector}
 \usage{
-array(x, type)
+array(x, type = NULL)
 }
 \arguments{
 \item{x}{R object}
 
-\item{type}{currently ignored}
+\item{type}{Explicit \link[=arrow__DataType]{type}, or NULL (the default) to infer from the data}
 }
 \description{
 create an \link[=arrow__Array]{arrow::Array} from an R vector
