+++ r/tests/testthat/test-Array.R
@@ -285,12 +285,6 @@ test_that("support for NaN (ARROW-3615)", {
   expect_equal(y$null_count, 1L)
 })
 
-test_that("array ignores the type argument (ARROW-3784)", {
-  a <- expect_warning(array(1:10, type = int16()))
-  b <- array(1:10)
-  expect_equal(a, b)
-})
-
 test_that("integer types casts (ARROW-3741)", {
   a <- array(c(1:10, NA))
   a_int8 <- a$cast(int8())
